Add initial post around ad changes
This is a fleshed out version of my thinking,
and would love feedback on the ideas.
In particular,
does it logically make sense?

There will be some code changes required to update this,
but I'm considering this documentation driven dev.
So once we agree on a path forward,
I can implement it.