Add initial 2017 stats post.
This removes a lot of the commentary that was standard in prior years,
and just tries to get the stats out.

I don't think many people read the commentary,
so I think just having a history of the stats is better.
We can provide commentary in other formats that people are more likely to read.