Embedding Content From Your Documentation, blog post
We have covered most of this feature in our guide at
https://docs.readthedocs.io/en/stable/guides/embedding-content.html, so I don't
think there is too much to add here. However, I think the blog is the right
place to promote this feature and share a little of its story.

I'd like to have an example of in-app UI tooltip, but the tool we built does not
look nice and I don't think we want to promote that as is.