Note about ``build.jobs`` on Jupyter Book post
There is no need to tell people that `_config.yml` and `conf.py` need to be in
sync and pushed to the repository anymore since this is not required. Users can
use `build.jobs.pre_build` for this instead and generate the `conf.py` on the
fly when building on Read the Docs.