Post: `build.commands`
Initial content for a simple post announcing `build.commands` in beta.
It does not go too deep into it on purpose. The idea is just to announce there
is a new config key `build.commands` and how to use it.

There is going to be another post in the future that talks about the
integrations with all the platform features.