Post: remove support for Sphinx 1.x
Small changelog blog post to communicate we are deprecating installing Sphinx
1.x automatically for projects created before Oct 20, 2020.
Besides, the blog post briefly describes how to create a `requirements.txt`
file and how to update the `.readthedocs.yaml` to install them.

This is aligned with the deprecation of building without a config file.
We are using the same timeline for both.

The code removal of this feature flag is done at
https://github.com/readthedocs/readthedocs.org/pull/10365