Post: deprecation of building without config file (#219)
* Post: deprecation of building without config file

Initial blog post to point users to when talking about the deprecation of
"building without a configuration file".

I'm not mentioning a specific date on purpose here because we don't know yet.
However, I'm mentioning Sept 2023 because our goal is to match with the drop of
OpenSSL 1.1.1 support on `urllib`. We can update the post once we know this
exact date and also write a new pretty small blog post with an update of the
exact date later if we want to, but may not be necessary tho.

Once this blog post is published, we can update some links in the application to
point users to this post also.

Note that I wrote a "suggested configuration file" and alos describe some of the
differences, so users can dig a little more in case they start getting some
errors. I don't want to give them just a blob of YAML without guiding the a
little in the process.

* Bold the date

* Typo

* Typo

* Feedback

- add sections to split the content
- expand the bullets list to mention the dashboard settings and the
  corresponding YAML config key
- fix some typos

* Minor changes

* Ignore .direnv to avoid failing build locally

* Apply suggestions from code review

Co-authored-by: Anthony <aj@ohess.org>

* Minor .rst style changes

---------

Co-authored-by: Anthony <aj@ohess.org>